# [dynamo] exclude reset_rng_state when measure timing